sdk: Run the build as root
There's really no point in having an intermedia user for running the
build: we control the build environment, and we don't have an outside
system to mess up.

Since jhbuild bombs out when running as root, we patch it and add an
escape hatch for that via an environment variable.

After that, it's only a matter of setting up the appropriate paths when
creating the Docker image.